Change behavior of BigIntegerUtil to allow setBits to set leading 0s
In previous versions the setBits method took an instance of BigInteger but that did not
allow it to set a range of bits starting with 0. Now it takes a binary string and this
becomes possible. The tests have been updated to reflect the change and classes using
this functionnality have been updated too.

Closes #1
Closes #2