Silently shrink parameters in RFC4122 UUID
Previously an exception was throwed if parameters were larger
than expected. Now a bitwise and is used to reduce them to the
accepted size. It is thus not required to generate all versions
in the constructor to check for consistency, they are generated
just when required. Tests have been updated accordingly.