Group several methods to generate a random big integer
Two methods were previously defined, one in the test class for
RFC4122 UUIDs and another in the random RFC4122 UUID generator.
A new one has been created in BigIntegerUtil and is now used by
both the generator and the test class. A new test has also been
added for this new functionnality.

Closes #10