Add a script for manually checking bit distribution in UUIDs
This script is for manual use only. It is used to look at the
representation of bits in generated UUIDs. This can be used to
ensure that no bit is missed in computation, that could lead to
a bit always being at 0 or 1, and to ensure that version and
layout are actually set in RFC4122 UUIDs.